added section headings in pdf
elements are now annotated with classNames that contain the section identifier loinc code. With that at the time of pdf creation, each element can be associated with its section class. The information for the sectionsDisplay values in the pdf rendering process is taking directly from the config file. The EmptyComponentRenderer is also rendered if necessary