[Feature #139] pdf export (#163)
* broken export pdf

* html2canvas

* pdf almost ready yay

* add sleep and opening of accord

* 139 fix pdf renderer to work with safari and chrome

switch out html2canvas with html-to-image and use jpeg instead of png; add header and footer; add loading animation;

* added section headings in pdf

elements are now annotated with classNames that contain the section identifier loinc code. With that at the time of pdf creation, each element can be associated with its section class. The information for the sectionsDisplay values in the pdf rendering process is taking directly from the config file. The EmptyComponentRenderer is also rendered if necessary

* might need to revert this commit

* bester pdf wo gibt

* remove seconds

* align header to the left

* [Feature #139] refactoring

---------

Co-authored-by: Anja <anja.schwab939@gmail.com>
Co-authored-by: msmidt <maxismidt@icloud.com>